Move AutofillDeclineCounter to impl module
